requirements.txt.txt
aiogram
requests
python-dotenv
aiohttp